move FE to model utils, split in feature distillation plugin and feature replay strategy
